requirements.txt
attrs==25.4.0
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
flasgger==0.9.7.1
Flask==2.2.5
flask-cors==6.0.1
Flask-SQLAlchemy==3.0.3
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwt==1.4.0
MarkupSafe==3.0.3
mistune==3.1.4
packaging==25.0
pycparser==2.23
python-dotenv==1.0.0
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
